Add code for generating datasets
So far we have the basic scaffolding to go from simple information about a Pokemon the the values we need for our convex optimizer. We still need to discuss what exactly we want to vary based on dataset, vs what we want to keep fixed (i.e. is the pokemon species the same, and we look at usage rates? Or do we vary species and take the max offensive numbers for each species? Do we vary attacker species or defender species? Do we take a max across multiple species?)